microservices/topic-prioritization-service/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.14.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.23.3
python-json-logger==2.0.7
apscheduler==3.10.4 